Minor cleanup.
git-svn-id: svn+ssh://software.sandia.gov/svn/pecos/trunk@472 ccfab4c5-ad45-0410-a689-f63b6aebdeca

+++ src/LHSDriver.hpp
@@ -125,8 +125,8 @@ private:
   bool reportFlag;
 
   /// for honoring advance_seed_sequence() calls
-  int allowSeedAdvance; // bit 1 = first-time flag
-		        // bit 2 = allow repeated seed update
+  short allowSeedAdvance; // bit 1 = first-time flag
+		          // bit 2 = allow repeated seed update
 };
